# Faster campaigns test. Part 4. Reorganise tests.